refactor(ui): extract shared constants
