+++ apps/web/src/components/nav-projects.tsx
@@ -20,8 +20,8 @@ import {
   SidebarMenuAction,
   SidebarMenuButton,
   SidebarMenuItem,
-  useSidebar,
 } from "@/components/ui/sidebar"
+import { useSidebar } from "@/components/ui/use-sidebar"
 
 export function NavProjects({
   projects,
